+++ CHANGELOG.md
@@ -17,6 +17,9 @@ Full documentation for hipSOLVER is available at the [hipSOLVER Documentation](h
   `BUILD_HIPSPARSE_TESTS` CMake option to enable tests for the hipsolverSp API (on by default).
 
 ### Deprecated
+- The Fortran bindings provided in `hipsolver_module.f90` have been deprecated.
+  The Fortran bindings provided by the hipfort project are recommended instead.
+
 ### Removed
 ### Fixed
 ### Known Issues